Rename package to com.spelliar and update app name
