+++ app/src/commonMain/kotlin/com/spelliar/design/navigation/Routes.kt
@@ -1,0 +1,8 @@
+package com.spelliar.design.navigation
+
+import kotlinx.serialization.Serializable
+
+sealed interface Routes {
+    @Serializable
+    data object Main : Routes
+}
