--- app/src/commonMain/kotlin/ru/tsum/design/navigation/Routes.kt
@@ -1,8 +1,0 @@
-package ru.tsum.design.navigation
-
-import kotlinx.serialization.Serializable
-
-sealed interface Routes {
-    @Serializable
-    data object Main : Routes
-}
