--- app/src/commonMain/kotlin/ru/tsum/design/ui/MainView.kt
@@ -1,20 +1,0 @@
-package ru.tsum.design.ui
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import ru.tsum.design.navigation.Routes
-
-
-@Composable
-fun MainView(onNavigation: (Routes) -> Unit) {
-    MainViewContent(
-        onNavigation = onNavigation
-    )
-}
-
-@Composable
-fun MainViewContent(
-    onNavigation: (Routes) -> Unit
-) {
-    Text("MainView")
-}
