+++ app/src/iosMain/kotlin/com/spelliar/design/ui/MainViewController.kt
@@ -1,0 +1,7 @@
+package com.spelliar.design.ui
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun MainViewController() = ComposeUIViewController {
+    Navigation()
+}
