--- app/src/iosMain/kotlin/ru/tsum/design/ui/MainViewController.kt
@@ -1,7 +1,0 @@
-package ru.tsum.design.ui
-
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() = ComposeUIViewController {
-    Navigation()
-}
