+++ app/src/main/java/com/spelliar/design/MainActivity.kt
@@ -1,0 +1,17 @@
+package com.spelliar.design
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.spelliar.design.ui.Navigation
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            Navigation()
+        }
+    }
+}
