+++ app/src/main/java/com/spelliar/design/ui/MainViewContentPreview.kt
@@ -1,0 +1,12 @@
+package com.spelliar.design.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview
+@Composable
+private fun MainViewContentPreview() {
+    MainViewContent(
+        onNavigation = {}
+    )
+}
