--- app/src/main/java/ru/tsum/design/MainActivity.kt
@@ -1,17 +1,0 @@
-package ru.tsum.design
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import ru.tsum.design.ui.Navigation
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            Navigation()
-        }
-    }
-}
