--- app/src/main/java/ru/tsum/design/ui/MainViewContentPreview.kt
@@ -1,12 +1,0 @@
-package ru.tsum.design.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-
-@Preview
-@Composable
-private fun MainViewContentPreview() {
-    MainViewContent(
-        onNavigation = {}
-    )
-}
