+++ app/src/wasmJsMain/kotlin/com/spelliar/conventionplugins/demo/compose/app/main.kt
@@ -1,0 +1,13 @@
+package com.spelliar.conventionplugins.demo.compose.app
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.ComposeViewport
+import kotlinx.browser.document
+import com.spelliar.design.ui.Navigation
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    ComposeViewport(document.body!!) {
+        Navigation()
+    }
+}
