--- app/src/wasmJsMain/kotlin/ru/tsum/conventionplugins/demo/compose/app/main.kt
@@ -1,13 +1,0 @@
-package ru.tsum.conventionplugins.demo.compose.app
-
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.ComposeViewport
-import kotlinx.browser.document
-import ru.tsum.design.ui.Navigation
-
-@OptIn(ExperimentalComposeUiApi::class)
-fun main() {
-    ComposeViewport(document.body!!) {
-        Navigation()
-    }
-}
